Add custom logger
It is generally bad practice to leave console logs in the code
which could be publicly displayed to end users. Setting up our own
custom logger to prevent that situation and storing the log files
locally. New log files are created daily and a maximum of 14 log
files will be stored.